+++ RELEASE_NOTES.md
@@ -6,10 +6,10 @@
 <!------------------------------------------------------------------------------------------------->
 
 ### What's New
-/
+- Add estimation of CPU latencies for benchmark results (per iter of benchmark) & show after running benchmark
 
 ### Improvements
-/
+- Add CPU frequency to benchmark system_info.
 
 ### Bug Fixes
 /